requirements.txt
# scipy==1.0.0
Pillow==5.0.0 
steganography==0.1.1
pyexiftool==0.1.1
Flask==1.0.2 
gunicorn
pycrypto==2.6.1
scipy==1.3.0   